Floating point internal array, .HDR output, several bug fixes.

+++ RayTracing/vec3.h
@@ -14,22 +14,22 @@ class vec3 {
 public:
 
 	vec3() {}
-	vec3(float e0, float e1, float e2) { e[0] = e0; e[1] = e1; e[2] = e2; }
+	vec3(double e0, double e1, double e2) { e[0] = e0; e[1] = e1; e[2] = e2; }
 
 	//Can use x, y, z or r, g, b to return stored values - vec3 can be used to store either.
-	inline float x() const { return e[0]; }
-	inline float y() const { return e[1]; }
-	inline float z() const { return e[2]; }
-	inline float r() const { return e[0]; }
-	inline float g() const { return e[1]; }
-	inline float b() const { return e[2]; }
+	inline double x() const { return e[0]; }
+	inline double y() const { return e[1]; }
+	inline double z() const { return e[2]; }
+	inline double r() const { return e[0]; }
+	inline double g() const { return e[1]; }
+	inline double b() const { return e[2]; }
 
 
 	//Defining vector operators
 	inline const vec3& operator+() const { return *this; }
 	inline vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-	inline float operator[](int i) const { return e[i]; }
-	inline float& operator[](int i) { return e[i]; };
+	inline double operator[](int i) const { return e[i]; }
+	inline double& operator[](int i) { return e[i]; };
 
 	inline vec3& operator+=(const vec3 &v2);
 	inline vec3& operator-=(const vec3 &v2);
@@ -38,18 +38,18 @@ public:
 	inline vec3& operator*=(const float t);
 	inline vec3& operator/=(const float t);
 
-	inline float length() const {
+	inline double length() const {
 		return sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
 	}
 
 
-	inline float squared_length() const {
+	inline double squared_length() const {
 		return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
 	}
 
 	inline void make_unit_vector();
 
-	float e[3];
+	double e[3];
 };
 
 inline std::istream& operator>>(std::istream &is, vec3 &t) {
